#2032: renamed rust_project -> rust
