+++ eo-maven-plugin/src/main/java/org/eolang/maven/BinarizeParseMojo.java
@@ -46,9 +46,9 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 import org.eolang.maven.footprint.FtDefault;
-import org.eolang.maven.rust_project.Names;
-import org.eolang.maven.rust_project.Native;
-import org.eolang.maven.rust_project.Project;
+import org.eolang.maven.rust.Names;
+import org.eolang.maven.rust.Native;
+import org.eolang.maven.rust.Project;
 import org.eolang.maven.tojos.ForeignTojo;
 import org.eolang.maven.util.Home;
 import org.eolang.parser.ParsingTrain;
