+++ eo-maven-plugin/src/main/java/org/eolang/maven/rust/Names.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven.rust_project;
+package org.eolang.maven.rust;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
