+++ eo-maven-plugin/src/test/java/org/eolang/maven/rust/NativeTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven.rust_project;
+package org.eolang.maven.rust;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
